# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JAC Rental Booking Service API",
        "description": "Standalone microservice handling all rental property booking operations for the JAC platform. Authentication uses Bearer tokens issued by the JAC monolith (Laravel Passport).",
        "contact": {
            "email": "dev@jac.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "Current Environment"
        }
    ],
    "paths": {
...
        "/api/v1/units/{id}": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Get unit details",
                "description": "Returns full details for a specific unit including amenities and price rules.",
                "operationId": "85d6a4febe60b56171f40de3129bd949",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnitDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Unit not found"
                    }
                }
            }
        },
        "/api/v1/units/{id}/prices": {
            "get": {
                "tags": [
                    "Units"
                ],
                "summary": "Get custom price schedule for a unit",
                "description": "Returns override/seasonal price records for a given unit within a date range.",
                "operationId": "04732472c5f21d9cfe8aa3bd2e5477d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-09-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-09-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of price records in the requested period",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PriceRecord"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Units"
                ],
                "summary": "Set a custom price for a unit on a specific date",
                "description": "Creates or updates an override price for a specific date. Used by org managers/staff.",
                "operationId": "eba4c9ab0b9c7bbd397dd4bde0930bfb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "price"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-25"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 350
                                    },
                                    "currency": {
                                        "description": "SAR | USD | YER_S | YER_N",
                                        "type": "string",
                                        "example": "SAR"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Price set successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Price updated successfully"
                                        },
                                        "price": {
                                            "$ref": "#/components/schemas/PriceRecord"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/webhooks/payment-complete": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Payment gateway completion webhook",
                "description": "Called by monolith payment system after payment succeeds or fails.",
                "operationId": "997cb0b5cd388a93b334c65ff33eb425",
                "parameters": [
                    {
                        "name": "X-Webhook-Signature",
                        "in": "header",
                        "description": "HMAC-SHA256 of payload",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hold_token",
                                    "payment_reference",
                                    "status"
                                ],
                                "properties": {
                                    "hold_token": {
                                        "type": "string",
                                        "example": "HOLD-ABCDE12345XY"
                                    },
                                    "payment_reference": {
                                        "type": "string",
                                        "example": "PAY-9988776655"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "paid"
                                    },
                                    "gateway": {
                                        "type": "string",
                                        "example": "kuraimi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook processed — booking created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "booking_id": {
                                            "type": "integer",
                                            "example": 88
                                        },
                                        "reference_no": {
                                            "type": "string",
                                            "example": "BK-ABCD1234"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Hold not found or payment not paid"
                    },
                    "401": {
                        "description": "Invalid HMAC signature"
                    },
                    "500": {
                        "description": "Internal processing error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PriceRecord": {
                "description": "A price rule for a unit",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "priceable_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "priceable_type": {
                        "type": "string",
                        "example": "App\\Models\\Unit"
                    },
                    "price_type": {
                        "type": "string",
                        "example": "default"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "price_sar": {
                        "type": "number",
                        "format": "float",
                        "example": 350
                    },
                    "price_usd": {
                        "type": "number",
                        "format": "float",
                        "example": 93
                    },
                    "price_yer_s": {
                        "type": "number",
                        "format": "float",
                        "example": 87500
                    },
                    "price_yer_n": {
                        "type": "number",
                        "format": "float",
                        "example": 90000
                    }
                },
                "type": "object"
            },
...
            "PropertySummary": {
                "description": "Compact property representation returned in list views",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sea View Villa"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "avg_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "rating_count": {
                        "type": "integer",
                        "example": 12
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "star_rating": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "logo": {
                        "type": "string",
                        "example": "properties/villa1.jpg",
                        "nullable": true
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "UnitDetail": {
                "description": "Full unit details with amenities and price rules",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "property_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Chalet 101"
                    },
                    "pricing_mode": {
                        "type": "string",
                        "example": "nightly"
                    },
                    "max_guests": {
                        "type": "integer",
                        "example": 6
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "amenities": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "name": {
                                    "type": "string",
                                    "example": "WiFi"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "property": {
                        "$ref": "#/components/schemas/PropertySummary"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Passport Bearer token issued by the JAC monolith. Obtain via the monolith /api/v1/auth/login endpoint.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}